server/requirements.txt
acres==0.1.0
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
anndata==0.11.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
array_api_compat==1.9.1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.9.0
cachetools==5.5.0
cat-python==1.1.0
cat-sdk==3.1.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cmake==3.31.1
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
effdet==0.4.1
einops==0.8.0
emoji==2.14.0
eofs==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.0
faiss-cpu==1.9.0.post1
fastapi==0.115.0
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
fonttools==4.55.0
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface==0.0.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json-parser==1.2.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-nvidia-ai-endpoints==0.3.5
langchain-openai==0.2.2
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.132
layoutparser==0.3.4
libmagic==1.0
llama-index-core==0.12.3
llama-index-postprocessor-nvidia-rerank==0.4.0
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.14.2
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.1
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.20.1
openai==1.51.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.21.0
pikepdf==9.4.2
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.3
plotly==5.24.1
polars==1.15.0
portalocker==3.0.0
proto-plus==1.25.0
protobuf==5.28.3
prov==2.0.1
psutil==6.1.0
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.14
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.5.4
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.17
python-oxmsg==0.0.1
pytils==0.4.1
pytz==2024.2
pywin32==308
pyxnat==1.6.2
PyYAML==6.0.2
RapidFuzz==3.10.1
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
setuptools-scm==8.1.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
streamlit==1.40.2
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
toml==0.10.2
tools==0.1.9
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.5
traits==6.4.3
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.8
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
urllib3==2.2.3
uvicorn==0.31.1
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.17.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.13.1